# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billing application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── Dashboard ──
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary with totals
      parameters:
        - name: financialYear
          in: query
          schema:
            type: string
            description: "Financial year e.g. 2025-2026"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceSummary"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - name: financialYear
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/top-customers:
    get:
      operationId: getTopCustomers
      tags: [dashboard]
      summary: Get top customers by revenue
      parameters:
        - name: financialYear
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCustomer"
  /dashboard/monthly-sales:
    get:
      operationId: getMonthlySales
      tags: [dashboard]
      summary: Get monthly sales breakdown
      parameters:
        - name: financialYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Monthly sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlySales"

  # ── Reports ──
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      required: [productId, productName, quantity, unitPrice, gstPercent, gstAmount, totalAmount]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        costPrice:
          type: number
        gstPercent:
          type: number
        gstAmount:
          type: number
        discountPercent:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        profit:
          type: number
    CreateInvoiceItem:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        discountPercent:
          type: number
          default: 0
    Invoice:
      type: object
      required: [id, invoiceNumber, date, items, subtotal, totalGst, totalDiscount, grandTotal, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        totalGst:
          type: number
        totalDiscount:
          type: number
        grandTotal:
          type: number
        totalProfit:
          type: number
        notes:
          type: string
        createdAt:
          type: string
    CreateInvoiceBody:
      type: object
      required: [items, date]
      properties:
        customerId:
          type: integer
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceItem"
        notes:
          type: string
        overallDiscountPercent:
          type: number
          default: 0
    InvoiceSummary:
      type: object
      required: [id, invoiceNumber, date, grandTotal, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerName:
          type: string
        date:
          type: string
        grandTotal:
          type: number
        totalProfit:
          type: number
        itemCount:
          type: integer
        createdAt:
          type: string
    InvoiceListResponse:
      type: object
      required: [invoices, total, page, limit]
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DashboardSummary:
      type: object
      required: [totalSales, totalProfit, totalInvoices, totalCustomers, totalProducts]
      properties:
        totalSales:
          type: number
        totalProfit:
          type: number
        totalGst:
          type: number
        totalInvoices:
          type: integer
        totalCustomers:
          type: integer
        totalProducts:
          type: integer
        financialYear:
          type: string
    TopProduct:
      type: object
      required: [productId, productName, totalQuantity, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
    TopCustomer:
      type: object
      required: [customerId, customerName, totalPurchases, invoiceCount]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        totalPurchases:
          type: number
        invoiceCount:
          type: integer
    MonthlySales:
      type: object
      required: [month, year, totalSales, totalProfit, invoiceCount]
      properties:
        month:
          type: integer
        year:
          type: integer
        monthName:
          type: string
        totalSales:
          type: number
        totalProfit:
          type: number
        totalGst:
          type: number
        invoiceCount:
          type: integer
...
    DeleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string