# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceItem } from "./createInvoiceItem";

export interface CreateInvoiceBody {
  customerId?: number;
  date: string;
  items: CreateInvoiceItem[];
  notes?: string;
  overallDiscountPercent?: number;
}