# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSales: number;
  totalProfit: number;
  totalGst?: number;
  totalInvoices: number;
  totalCustomers: number;
  totalProducts: number;
  financialYear?: string;
}